backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
numpy
google-genai